arreglos de la vista y clases afiliados

+++ src/Clinica Frba/Clases/Afiliado.cs
@@ -12,8 +12,8 @@ namespace Clinica_Frba.Clases
 {
     class Afiliado: Persona
     {
-        public int Codigo_Persona { get; set; }
-        public int Numero_Familiar { get; set; }
+        public decimal Codigo_Persona { get; set; }
+        public decimal Numero_Familiar { get; set; }
         public decimal Numero_Grupo { get; set; }
         public string Estado_Civil { get; set; }
         public int Cantidad_Hijos { get; set; }
